requirements.txt
# external requirements
click~=8.1.7
coverage==7.3.2
flake8==6.1.0
python-dotenv>=0.5.1
torch==2.0.1
lightning==2.1.2
torchvision==0.15.2
numpy~=1.26.2
matplotlib~=3.8.2
tqdm~=4.66.1
torchmetrics==1.2.0
hydra-core==1.3.2
awscli==1.30.2
boto3==1.29.2
botocore==1.32.2
pip~=22.3.1
attrs~=23.1.0
wheel~=0.38.4
docutils~=0.16
Jinja2~=3.1.2
MarkupSafe~=2.1.3
werkzeug~=3.0.1
blinker~=1.7.0
itsdangerous~=2.1.2
setuptools~=65.5.1
sympy~=1.12
mpmath~=1.3.0
packaging~=23.2
colorama~=0.4.4
urllib3~=2.0.7
fsspec~=2023.10.0
requests~=2.31.0
aiohttp~=3.9.0
yarl~=1.9.3
certifi~=2023.11.17
networkx~=3.2.1
idna~=3.4
PyYAML~=6.0.1
pyparsing~=3.1.1
flask~=3.0.0
omegaconf~=2.3.0
Pillow~=9.4.0
wandb
